fixato creazione dei file config.json e requerement.txt nella folder del bot
