+++ StreamingCommunity/Api/Site/altadefinizionegratis/film.py
@@ -8,9 +8,7 @@ from StreamingCommunity.Util.console import console
 from StreamingCommunity.Util.os import os_manager
 from StreamingCommunity.Util.message import start_message
 from StreamingCommunity.Lib.Downloader import HLS_Downloader
-from StreamingCommunity.TelegramHelp.telegram_bot import get_bot_instance
-from StreamingCommunity.TelegramHelp.session import get_session, updateScriptId, deleteScriptId
-
+from StreamingCommunity.TelegramHelp.telegram_bot import TelegramSession, get_bot_instance
 
 # Logic class
 from StreamingCommunity.Api.Template.Class.SearchType import MediaItem
@@ -38,17 +36,17 @@ def download_film(select_title: MediaItem) -> str:
     if TELEGRAM_BOT:
         bot = get_bot_instance()
         bot.send_message(f"Download in corso:\n{select_title.name}", None)
-    
+
         # Get script_id
-        script_id = get_session()
+        script_id = TelegramSession.get_session()
         if script_id != "unknown":
-            updateScriptId(script_id, select_title.name)
+            TelegramSession.updateScriptId(script_id, select_title.name)
 
     # Start message and display film information
     start_message()
     console.print(f"[yellow]Download:  [red]{select_title.name} \n")
     console.print(f"[cyan]You can safely stop the download with [bold]Ctrl+c[bold] [cyan] \n")
-    
+
     # Set domain and media ID for the video source
     video_source = VideoSource(select_title.url)
 
@@ -61,17 +59,17 @@ def download_film(select_title: MediaItem) -> str:
 
     # Download the film using the m3u8 playlist, and output filename
     r_proc = HLS_Downloader(
-        m3u8_url=master_playlist, 
+        m3u8_url=master_playlist,
         output_path=os.path.join(mp4_path, title_name)
     ).start()
 
     if TELEGRAM_BOT:
-      
+
         # Delete script_id
-        script_id = get_session()
+        script_id = TelegramSession.get_session()
         if script_id != "unknown":
-            deleteScriptId(script_id)
-          
+            TelegramSession.deleteScriptId(script_id)
+
     if "error" in r_proc.keys():
         try:
             os.remove(r_proc['path'])
